refactor to smaller functions
